requirements.txt
click==7.1.2
Columnar==1.3.1
docopt==0.6.2
Flask>=1
flask-cors>=3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
obd==0.7.1
Pint==0.7.2
pyserial==3.4
PyYAML==5.3.1
toolz==0.10.0
wcwidth==0.2.4
Werkzeug==1.0.1
influxdb==5.3.0